chore: move model pricing out of settings
