+++ frontend/src/container/AIObservability/ModelPricing/AIObservabilityModelPricing.tsx
