+++ frontend/src/container/AIObservability/ModelPricing/ModelCostTabPanel/components/ModelCostDrawer/components/PatternEditor/index.ts
