+++ frontend/src/container/AIObservability/ModelPricing/ModelCostTabPanel/components/ModelCostsTable/ModelCostsTable.module.scss
