+++ frontend/src/container/AIObservability/ModelPricing/ModelCostTabPanel/components/ModelCostsTable/ModelCostsTable.tsx
