+++ frontend/src/container/AIObservability/ModelPricing/ModelCostTabPanel/components/ModelCostsTable/TableConfig/index.ts
