+++ frontend/src/container/AIObservability/ModelPricing/ModelCostTabPanel/components/ModelCostsTable/TableConfig/table.config.tsx
