+++ frontend/src/container/AIObservability/ModelPricing/UnpricedModelsTab/components/UnpricedModelsTable/TableConfig/index.ts
