+++ frontend/src/container/AIObservability/ModelPricing/__tests__/fixtures.ts
